+++ common/src/main/java/dev/ftb/mods/ftbultimine/api/restriction/RestrictionHandler.java
@@ -1,5 +1,6 @@
 package dev.ftb.mods.ftbultimine.api.restriction;
 
+import dev.ftb.mods.ftbultimine.api.util.CanUltimineResult;
 import net.minecraft.world.entity.player.Player;
 
 /**
@@ -9,8 +10,17 @@ import net.minecraft.world.entity.player.Player;
 public interface RestrictionHandler {
     /**
      * Is the player allowed to do ultimining at this time?
+     *
      * @param player the player being checked
      * @return false to prevent ultimining, true to allow (note that if any handlers are registered, <em>all</em> must return true for ultimining to be allowed)
      */
     boolean canUltimine(Player player);
+
+    /**
+     * This is used to get the reason for the restriction to display to the player.
+     * @return translatable key with the reason Ultimine is being blocked.
+     */
+    default CanUltimineResult ultimineBlockReason(Player player) {
+        return CanUltimineResult.OTHER_RESTRICTION;
+    }
 }